+++ crates/guest-write-file/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [0.1.109](https://github.com/vm0-ai/vm0/compare/guest-write-file-v0.1.108...guest-write-file-v0.1.109) (2026-08-30)
+
+## [0.1.108](https://github.com/vm0-ai/vm0/compare/guest-write-file-v0.1.107...guest-write-file-v0.1.108) (2026-08-30)
+
 ## [0.1.107](https://github.com/vm0-ai/vm0/compare/guest-write-file-v0.1.106...guest-write-file-v0.1.107) (2026-08-29)
 
 ## [0.1.106](https://github.com/vm0-ai/vm0/compare/guest-write-file-v0.1.105...guest-write-file-v0.1.106) (2026-08-28)